ArrayExamples Update
For reversed, fixed issue where returns the incorrect array and made sure to update the correct array. Also, for reverseInPlace, made sure to not undo changes (through arr.length / 2) and save a temp of arr[i] to update the last half of the array. Also, updated the tests to ensure that they work as intended.